cli: refactor to align with katautils package
refactor the cli codes which can be shared with shimv2.
