+++ cli/utils.go
@@ -15,8 +15,7 @@ import (
 )
 
 const (
-	unknown     = "<<unknown>>"
-	k8sEmptyDir = "kubernetes.io~empty-dir"
+	unknown = "<<unknown>>"
 )
 
 // variables to allow tests to modify the values
@@ -34,26 +33,6 @@ func fileExists(path string) bool {
 	}
 
 	return true
-}
-
-// IsEphemeralStorage returns true if the given path
-// to the storage belongs to kubernetes ephemeral storage
-//
-// This method depends on a specific path used by k8s
-// to detect if it's of type ephemeral. As of now,
-// this is a very k8s specific solution that works
-// but in future there should be a better way for this
-// method to determine if the path is for ephemeral
-// volume type
-func IsEphemeralStorage(path string) bool {
-	splitSourceSlice := strings.Split(path, "/")
-	if len(splitSourceSlice) > 1 {
-		storageType := splitSourceSlice[len(splitSourceSlice)-2]
-		if storageType == k8sEmptyDir {
-			return true
-		}
-	}
-	return false
 }
 
 func getKernelVersion() (string, error) {
